requirements.txt
Django==2.2.13
djangorestframework==3.11.0
gunicorn==20.0.4
messenger-bot-library==2.0.2b0
phonenumbers==8.12.5
psycopg2-binary==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
